Melhorias enums

+++ InfiniteRoleplay/Commands/Cellphone.cs
@@ -89,7 +89,7 @@ namespace InfiniteRoleplay.Commands
             if (numero == 5555555)
             {
                 Functions.EnviarMensagem(player, TipoMensagem.Nenhum, "!{#e6a250}" + $"[CELULAR] Você está ligando para {p.ObterNomeContato(numero)}.");
-                if (Global.PersonagensOnline.Count(x => x.Emprego == (int)TipoEmprego.Taxista && x.IsEmTrabalho) == 0)
+                if (Global.PersonagensOnline.Count(x => x.Emprego == TipoEmprego.Taxista && x.IsEmTrabalho) == 0)
                 {
                     Functions.EnviarMensagem(player, TipoMensagem.Nenhum, "!{#F0E90D}" + $"[CELULAR] {p.ObterNomeContato(numero)} diz: Desculpe, não temos nenhum taxista em serviço no momento.");
                     return;
